TileMapEditor
Viewer에 현재 타일 오브젝트가 가지고 있는 타일맵 띄우기 까지 성공

+++ Project/Client/SE_AtlasView.cpp
@@ -54,7 +54,7 @@ void SE_AtlasView::SetAtlasTex(Ptr<CTexture> _Tex)
 
 	m_AtlasTex = _Tex;
 
-	m_WidthSize = m_AtlasTex->Width();
+	m_WidthSize = (float)m_AtlasTex->Width();
 }
 
 void SE_AtlasView::WheelCheck()
